trading_api/requirements.txt
aiogram==2.9.2
aiohttp==3.6.2
aioredis==1.3.1
amqp==2.6.0
APScheduler==3.6.3
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
base58==2.0.1
billiard==3.6.3.0
bitarray==1.2.2
celery==4.4.6
certifi==2020.6.20
chardet==3.0.4
cytoolz==0.10.1
Django==3.0.8
djangorestframework==3.11.0
eth-abi==2.1.1
eth-account==0.5.2
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.1.2
eth-typing==2.2.1
eth-utils==1.9.0
future==0.18.2
gunicorn==20.0.4
hexbytes==0.2.1
hiredis==1.1.0
idna==2.10
ipfshttpclient==0.6.0.post1
jsonschema==3.2.0
kombu==4.6.11
lru-dict==1.1.6
multiaddr==0.0.9
multidict==4.7.6
netaddr==0.8.0
parsimonious==0.8.1
protobuf==3.12.2
psycopg2-binary==2.8.5
pycryptodome==3.9.8
pyrsistent==0.16.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
rlp==1.2.0
six==1.15.0
sqlparse==0.3.1
toolz==0.10.0
tzlocal==2.1
urllib3==1.25.9
varint==1.0.2
vine==1.3.0
web3==5.11.1
websockets==8.1
yarl==1.4.2
